requirements.txt
-i https://mirrors.aliyun.com/pypi/simple
aiofiles==22.1.0
fastapi==0.87.0
PyYAML==6.0
requests==2.28.1
uvicorn==0.20.0
colorlog==6.7.0
joblib==1.2.0
ConcurrentLogHandler==0.9.1